Migrate all the things
